core: add get all receipts by hash helper
